updating

+++ zoodatasets/chunkdatasets.py
@@ -95,7 +95,7 @@ class ZooDataset(Dataset):
     def __getitem__(self, idx):
         if torch.is_tensor(idx):
             idx = idx.tolist()
-        weight = self.data[idx].to(torch.float32)
+        weight = self.data[idx]
         if self.transform:
             weight = self.transform(weight)
